+++ {{cookiecutter.github_repo_name}}/terraform/stacks/modules/kubernetes_ingress_clb/route53.tf
@@ -6,8 +6,8 @@
 #
 # usage: create DNS records for EKS cluster load balancer
 #------------------------------------------------------------------------------
-data "aws_route53_zone" "admin_domain" {
-  name = var.admin_domain
+data "aws_route53_zone" "services_subdomain" {
+  name = var.services_subdomain
 }
 # to-do: remove this declaration and refactor references below from
 # data.kubernetes_service.ingress_nginx_controller to
@@ -32,37 +32,13 @@ data "aws_route53_zone" "root_domain" {
   name = var.root_domain
 }
 
-resource "aws_route53_record" "root_naked" {
-  zone_id = data.aws_route53_zone.root_domain.id
-  name    = var.root_domain
-  type    = "A"
-
-  alias {
-    name                   = data.kubernetes_service.ingress_nginx_controller.status.0.load_balancer.0.ingress.0.hostname
-    zone_id                = data.aws_elb_hosted_zone_id.main.id
-    evaluate_target_health = true
-  }
-}
-
-resource "aws_route53_record" "root_wildcard" {
-  zone_id = data.aws_route53_zone.root_domain.id
-  name    = "*.${var.root_domain}"
-  type    = "A"
-
-  alias {
-    name                   = data.kubernetes_service.ingress_nginx_controller.status.0.load_balancer.0.ingress.0.hostname
-    zone_id                = data.aws_elb_hosted_zone_id.main.id
-    evaluate_target_health = true
-  }
-}
-
 # -------------------------------------------------------------------------------------
 # setup DNS for admin subdomain
 # -------------------------------------------------------------------------------------
 
 resource "aws_route53_record" "admin_naked" {
-  zone_id = data.aws_route53_zone.admin_domain.id
-  name    = var.admin_domain
+  zone_id = data.aws_route53_zone.services_subdomain.id
+  name    = var.services_subdomain
   type    = "A"
 
   alias {
@@ -73,8 +49,8 @@ resource "aws_route53_record" "admin_naked" {
 }
 
 resource "aws_route53_record" "admin_wildcard" {
-  zone_id = data.aws_route53_zone.admin_domain.id
-  name    = "*.${var.admin_domain}"
+  zone_id = data.aws_route53_zone.services_subdomain.id
+  name    = "*.${var.services_subdomain}"
   type    = "A"
 
   alias {