+++ {{cookiecutter.github_repo_name}}/terraform/stacks/modules/kubernetes_ingress_clb/variables.tf
@@ -7,7 +7,7 @@
 # usage: build an EKS cluster load balancer
 #------------------------------------------------------------------------------
 
-variable "admin_domain" {
+variable "services_subdomain" {
   type = string
 }