+++ {{cookiecutter.github_repo_name}}/terraform/stacks/modules/mysql/variables.tf
@@ -7,7 +7,7 @@
 # usage: create an RDS MySQL instance.
 #------------------------------------------------------------------------------
 
-variable "root_domain" {
+variable "services_subdomain" {
   type = string
 }
 variable "allocated_storage" {