+++ {{cookiecutter.github_repo_name}}/terraform/stacks/modules/vpc/route53.tf
@@ -2,16 +2,16 @@ data "aws_route53_zone" "root_domain" {
   name = var.root_domain
 }
 
-resource "aws_route53_zone" "admin_domain" {
-  name = var.admin_domain
+resource "aws_route53_zone" "services_subdomain" {
+  name = var.services_subdomain
   tags = var.tags
 }
 
 
 resource "aws_route53_record" "admin_domain_ns" {
   zone_id = data.aws_route53_zone.root_domain.zone_id
-  name    = aws_route53_zone.admin_domain.name
+  name    = aws_route53_zone.services_subdomain.name
   type    = "NS"
   ttl     = "30"
-  records = aws_route53_zone.admin_domain.name_servers
+  records = aws_route53_zone.services_subdomain.name_servers
 }